REFACTOR: split routes.utility_functions into those that are for file handling and those that are for job handling
